refactor(settings): removed AccountDataHOC
Because our assumptions about the data from gql have changed, we no longer need to worry about parts of the Account type needing a verified session, we can simplify the structure a bit.

In the future we can use `useQuery` in components to get the data they need instead of pushing down the `account` object.

This also introduces the `models` and `operations` folders for organizing types and read/write operations.